rearrange redux index - portfolio and isPreview
