+++ src/components/HeaderSidebar.tsx
@@ -1,15 +1,16 @@
 import React from "react";
 import styled from "styled-components";
-
 import { Link, useNavigate } from "react-router-dom";
 import { UserReducer } from "../reducers";
 import { signOut } from "firebase/auth";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faXmark } from "@fortawesome/free-solid-svg-icons";
+
 import { userLoading } from "../action/UserReducerAction";
 import { resumeLoading } from "../action/ResumeReducerAction";
 import { websiteLoading } from "../action/WebsiteReducerAction";
-import { portfolioLoading, setAlert } from "../action";
+import { portfolioLoading } from "../action/PortfolioReducerAction";
+import { setAlert } from "../action/IsPreviewReducerAction";
 import { useDispatch } from "react-redux";
 
 const Wrapper = styled.div<{ isSideBar: boolean }>`
