+++ src/pages/AllResumes/ResumeCard.tsx
@@ -3,7 +3,8 @@ import styled from "styled-components";
 import { Link, useNavigate } from "react-router-dom";
 import { DocumentData } from "firebase/firestore";
 import { useDispatch, useSelector } from "react-redux";
-import { portfolioLoading, setAlert } from "../../action";
+import { portfolioLoading } from "../../action/PortfolioReducerAction";
+import { setAlert } from "../../action/IsPreviewReducerAction";
 import { RootState } from "../../reducers";
 import firebase from "../../utilis/firebase";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
